+++ integration/testdata/helm/change_replicas.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+import sys
+import os
+stdin_data = sys.stdin.read()
+count = os.getenv("REPLICAS")
+
+print(stdin_data.replace("replicas: 1", "replicas: {}".format(count)))