+++ CHANGELOG.md
@@ -1,3 +1,123 @@
+# v2.9.0 Release - 11/07/2023
+**Linux amd64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.9.0/skaffold-linux-amd64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**Linux arm64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.9.0/skaffold-linux-arm64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**macOS amd64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.9.0/skaffold-darwin-amd64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**macOS arm64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.9.0/skaffold-darwin-arm64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**Windows**
+https://storage.googleapis.com/skaffold/releases/v2.9.0/skaffold-windows-amd64.exe
+
+**Docker image**
+`gcr.io/k8s-skaffold/skaffold:v2.9.0`
+
+Note: This release comes with a new config version, `v4beta8`. To upgrade your skaffold.yaml, use `skaffold fix`. If you choose not to upgrade, skaffold will auto-upgrade as best as it can.
+
+New Features and Additions:
+* chore: add new skip-unreachable-dirs to not error on init command when a dir can not be read [#9163](https://github.com/GoogleContainerTools/skaffold/pull/9163)
+* chore: add new config to control the pull behaviour for verify [#9150](https://github.com/GoogleContainerTools/skaffold/pull/9150)
+* chore: change custom actions pull logic, to check if an image exists locally first before triggering a pull [#9147](https://github.com/GoogleContainerTools/skaffold/pull/9147)
+
+Fixes:
+* fix: kpt force named "false" in schema [#9074](https://github.com/GoogleContainerTools/skaffold/pull/9074)
+
+Updates and Refactors:
+* chore: bump golang.org/x/net from 0.7.0 to 0.17.0 in /hack/tools [#9129](https://github.com/GoogleContainerTools/skaffold/pull/9129)
+* chore: bump golang.org/x/net from 0.7.0 to 0.17.0 in /examples/grpc-e2e-tests/service [#9130](https://github.com/GoogleContainerTools/skaffold/pull/9130)
+* chore: bump golang.org/x/net from 0.7.0 to 0.17.0 in /integration/examples/grpc-e2e-tests/cloud-spanner-bootstrap [#9131](https://github.com/GoogleContainerTools/skaffold/pull/9131)
+* chore: bump golang.org/x/net from 0.7.0 to 0.17.0 in /integration/examples/grpc-e2e-tests/service [#9128](https://github.com/GoogleContainerTools/skaffold/pull/9128)
+* chore: bump google.golang.org/grpc from 1.55.0 to 1.56.3 [#9156](https://github.com/GoogleContainerTools/skaffold/pull/9156)
+* chore: bump google.golang.org/grpc from 1.53.0 to 1.56.3 in /integration/examples/grpc-e2e-tests/service [#9154](https://github.com/GoogleContainerTools/skaffold/pull/9154)
+* chore: bump google.golang.org/grpc from 1.53.0 to 1.56.3 in /examples/grpc-e2e-tests/service [#9153](https://github.com/GoogleContainerTools/skaffold/pull/9153)
+* chore: bump google.golang.org/grpc from 1.53.0 to 1.56.3 in /examples/grpc-e2e-tests/cloud-spanner-bootstrap [#9152](https://github.com/GoogleContainerTools/skaffold/pull/9152)
+* chore: bump google.golang.org/grpc from 1.53.0 to 1.56.3 in /integration/examples/grpc-e2e-tests/cloud-spanner-bootstrap [#9155](https://github.com/GoogleContainerTools/skaffold/pull/9155)
+* chore: bump github/codeql-action from 2.22.4 to 2.22.5 [#9157](https://github.com/GoogleContainerTools/skaffold/pull/9157)
+* chore: bump ossf/scorecard-action from 2.3.0 to 2.3.1 [#9149](https://github.com/GoogleContainerTools/skaffold/pull/9149)
+* chore: bump schema version to v4beta8 [#9143](https://github.com/GoogleContainerTools/skaffold/pull/9143)
+* chore: bump peter-evans/create-or-update-comment from 3.0.2 to 3.1.0 [#9142](https://github.com/GoogleContainerTools/skaffold/pull/9142)
+* chore: bump github/codeql-action from 2.22.3 to 2.22.4 [#9146](https://github.com/GoogleContainerTools/skaffold/pull/9146)
+* chore: bump github/codeql-action from 2.22.2 to 2.22.3 [#9137](https://github.com/GoogleContainerTools/skaffold/pull/9137)
+* chore: bump golang.org/x/net from 0.7.0 to 0.17.0 in /examples/grpc-e2e-tests/cloud-spanner-bootstrap [#9132](https://github.com/GoogleContainerTools/skaffold/pull/9132)
+* chore: bump github/codeql-action from 2.22.1 to 2.22.2 [#9133](https://github.com/GoogleContainerTools/skaffold/pull/9133)
+* chore: bump ossf/scorecard-action from 2.2.0 to 2.3.0 [#9122](https://github.com/GoogleContainerTools/skaffold/pull/9122)
+* chore: bump github/codeql-action from 2.22.0 to 2.22.1 [#9123](https://github.com/GoogleContainerTools/skaffold/pull/9123)
+* chore: bump github/codeql-action from 2.21.9 to 2.22.0 [#9115](https://github.com/GoogleContainerTools/skaffold/pull/9115)
+* chore: bump flask from 2.3.3 to 3.0.0 in /integration/examples [#9107](https://github.com/GoogleContainerTools/skaffold/pull/9107)
+* chore: bump flask from 2.3.3 to 3.0.0 in /examples [#9106](https://github.com/GoogleContainerTools/skaffold/pull/9106)
+
+Docs, Test, and Release Updates:
+* docs: fix releaseNoteLink for v2.8.0 [#9125](https://github.com/GoogleContainerTools/skaffold/pull/9125)
+
+Huge thanks goes out to all of our contributors for this release:
+
+- Julian Tölle
+- Renzo Rojas
+- Zev Isert
+- dependabot[bot]
+
+# v2.8.0 Release - 10/03/2023
+**Linux amd64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.8.0/skaffold-linux-amd64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**Linux arm64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.8.0/skaffold-linux-arm64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**macOS amd64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.8.0/skaffold-darwin-amd64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**macOS arm64**
+`curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.8.0/skaffold-darwin-arm64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`
+
+**Windows**
+https://storage.googleapis.com/skaffold/releases/v2.8.0/skaffold-windows-amd64.exe
+
+**Docker image**
+`gcr.io/k8s-skaffold/skaffold:v2.8.0`
+
+Note: This release comes with a new config version, `v4beta7`. To upgrade your skaffold.yaml, use `skaffold fix`. If you choose not to upgrade, skaffold will auto-upgrade as best as it can.
+
+New Features and Additions:
+* feat: Support post-renderer for helm deployer. [#9100](https://github.com/GoogleContainerTools/skaffold/pull/9100)
+* feat: inject namespace from rendered manifests in post deploy hooks [#9090](https://github.com/GoogleContainerTools/skaffold/pull/9090)
+* feat: Add skaffold inspect command for adding config dependencies [#9072](https://github.com/GoogleContainerTools/skaffold/pull/9072)
+* feat: emit metrics for exec, verify and render [#9078](https://github.com/GoogleContainerTools/skaffold/pull/9078)
+* feat: Add global build pre- and post-hooks  [#9047](https://github.com/GoogleContainerTools/skaffold/pull/9047)
+* feat: allow specifying a remote config dependency from Google Cloud Storage [#9057](https://github.com/GoogleContainerTools/skaffold/pull/9057)
+
+Updates and Refactors:
+* chore: bump github/codeql-action from 2.21.8 to 2.21.9 [#9101](https://github.com/GoogleContainerTools/skaffold/pull/9101)
+* chore: bump github/codeql-action from 2.21.7 to 2.21.8 [#9097](https://github.com/GoogleContainerTools/skaffold/pull/9097)
+* chore: bump github/codeql-action from 2.21.6 to 2.21.7 [#9096](https://github.com/GoogleContainerTools/skaffold/pull/9096)
+* chore: add set docker host by current context [#9094](https://github.com/GoogleContainerTools/skaffold/pull/9094)
+* chore: bump github/codeql-action from 2.21.5 to 2.21.6 [#9093](https://github.com/GoogleContainerTools/skaffold/pull/9093)
+* chore: cherry-pick upgrade ko (#9043) to v2.7 [#9089](https://github.com/GoogleContainerTools/skaffold/pull/9089)
+* chore: verify should preserve job manifest envs [#9087](https://github.com/GoogleContainerTools/skaffold/pull/9087)
+* chore: bump actions/upload-artifact from 3.1.2 to 3.1.3 [#9075](https://github.com/GoogleContainerTools/skaffold/pull/9075)
+* chore: upgrade ko [#9043](https://github.com/GoogleContainerTools/skaffold/pull/9043)
+* chore: bump actions/checkout from 3 to 4 [#9067](https://github.com/GoogleContainerTools/skaffold/pull/9067)
+
+Docs, Test, and Release Updates:
+* docs: Fix document tutorials/skaffold-resource-selector.md [#9083](https://github.com/GoogleContainerTools/skaffold/pull/9083)
+* docs: add templatable field [#9088](https://github.com/GoogleContainerTools/skaffold/pull/9088)
+
+Huge thanks goes out to all of our contributors for this release:
+
+- Danilo Cianfrone
+- Matt Santa
+- Michael Plump
+- Renzo Rojas
+- Seita Uchimura
+- dependabot[bot]
+- ericzzzzzzz
+- guangwu
+- yosukei3108
+
 # v2.7.0 Release - 08/30/2023
 **Linux amd64**
 `curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/v2.7.0/skaffold-linux-amd64 && chmod +x skaffold && sudo mv skaffold /usr/local/bin`